fix(Paths): fixed paths helper import path
