+++ apps/web/src/app/[lang]/dho/[id]/@aside/(.)discussions/[slug]/page.tsx
@@ -4,7 +4,7 @@ import {
   getDiscussionBySlug,
 } from '@hypha-platform/graphql/rsc';
 import { SidePanel } from '../../_components/side-panel';
-import { Paths } from 'apps/web/src/app/constants';
+import { Paths } from '@hypha-platform/tools';
 import { Locale } from '@hypha-platform/i18n';
 
 type PageProps = {
