+++ tsconfig.base.json
@@ -10,7 +10,10 @@
     "importHelpers": true,
     "target": "es2015",
     "module": "esnext",
-    "lib": ["es2020", "dom"],
+    "lib": [
+      "es2020",
+      "dom"
+    ],
     "skipLibCheck": true,
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
@@ -18,27 +21,61 @@
       "@hypha-platform/authentication": [
         "packages/authentication/src/index.ts"
       ],
-      "@hypha-platform/cookie": ["packages/cookie/src/index.ts"],
-      "@hypha-platform/core": ["packages/core/src/index.ts"],
-      "@hypha-platform/epics": ["packages/epics/src/index.ts"],
-      "@hypha-platform/feature-flags": ["packages/feature-flags/src/index.ts"],
-      "@hypha-platform/graphql": ["packages/api/graphql/src/index.ts"],
-      "@hypha-platform/graphql/rsc": ["packages/api/graphql/src/rsc.ts"],
-      "@hypha-platform/i18n": ["packages/i18n/src/index.ts"],
-      "@hypha-platform/i18n/client": ["packages/i18n/src/client.ts"],
-      "@hypha-platform/json-rpc": ["packages/api/json-rpc/src/index.ts"],
-      "@hypha-platform/lib/utils": ["packages/ui-utils/src/index.ts"],
-      "@hypha-platform/model": ["packages/model/src/index.ts"],
+      "@hypha-platform/cookie": [
+        "packages/cookie/src/index.ts"
+      ],
+      "@hypha-platform/core": [
+        "packages/core/src/index.ts"
+      ],
+      "@hypha-platform/epics": [
+        "packages/epics/src/index.ts"
+      ],
+      "@hypha-platform/feature-flags": [
+        "packages/feature-flags/src/index.ts"
+      ],
+      "@hypha-platform/graphql": [
+        "packages/api/graphql/src/index.ts"
+      ],
+      "@hypha-platform/graphql/rsc": [
+        "packages/api/graphql/src/rsc.ts"
+      ],
+      "@hypha-platform/i18n": [
+        "packages/i18n/src/index.ts"
+      ],
+      "@hypha-platform/i18n/client": [
+        "packages/i18n/src/client.ts"
+      ],
+      "@hypha-platform/json-rpc": [
+        "packages/api/json-rpc/src/index.ts"
+      ],
+      "@hypha-platform/lib/utils": [
+        "packages/ui-utils/src/index.ts"
+      ],
+      "@hypha-platform/model": [
+        "packages/model/src/index.ts"
+      ],
       "@hypha-platform/storage-postgres": [
         "packages/storage-postgres/src/index.ts"
       ],
-      "@hypha-platform/ui": ["packages/ui/src/index.ts"],
-      "@hypha-platform/ui-utils": ["packages/ui-utils/src/index.ts"],
+      "@hypha-platform/ui": [
+        "packages/ui/src/index.ts"
+      ],
+      "@hypha-platform/ui-utils": [
+        "packages/ui-utils/src/index.ts"
+      ],
       "@hypha-platform/ui-utils/global.css": [
         "packages/ui-utils/src/global.css"
       ],
-      "@hypha-platform/ui/server": ["packages/ui/src/server.ts"]
+      "@hypha-platform/ui/server": [
+        "packages/ui/src/server.ts"
+      ],
+      "@hypha-platform/tools": [
+        "packages/tools/src/index.ts"
+      ]
     }
   },
-  "exclude": ["node_modules", "tmp"]
+  "exclude": [
+    "node_modules",
+    "tmp"
+  ]
 }
